refactor: replace cryptoStanceGradeElement with cryptoStanceGrade in DTSIPersonHeroCard components across various pages for consistency
