+++ src/components/app/pageHome/ca/index.tsx
@@ -1,5 +1,5 @@
 import { CryptoSupportHighlight } from '@/components/app/cryptoSupportHighlight'
-import { sortDTSIPersonDataTable } from '@/components/app/dtsiClientPersonDataTable/sortPeople'
+import { sortDTSIPersonDataTable } from '@/components/app/dtsiClientPersonDataTable/common/utils'
 import { DTSIPersonHeroCard } from '@/components/app/dtsiPersonHeroCard'
 import { DTSIPersonHeroCardRow } from '@/components/app/dtsiPersonHeroCard/dtsiPersonHeroCardRow'
 import { DTSIThumbsUpOrDownGrade } from '@/components/app/dtsiThumbsUpOrDownGrade'
@@ -113,7 +113,7 @@ export function CaPageHome({
             {highestScores.map(person => (
               <DTSIPersonHeroCard
                 countryCode={countryCode}
-                cryptoStanceGradeElement={<DTSIThumbsUpOrDownGrade person={person} />}
+                cryptoStanceGrade={DTSIThumbsUpOrDownGrade}
                 key={person.id}
                 person={person}
                 subheader="role-w-state"
@@ -129,7 +129,7 @@ export function CaPageHome({
             {lowestScores.map(person => (
               <DTSIPersonHeroCard
                 countryCode={countryCode}
-                cryptoStanceGradeElement={<DTSIThumbsUpOrDownGrade person={person} />}
+                cryptoStanceGrade={DTSIThumbsUpOrDownGrade}
                 key={person.id}
                 person={person}
                 subheader="role-w-state"
