+++ src/components/app/pageLocationKeyRaces/us/locationRaceSpecific/index.tsx
@@ -147,7 +147,7 @@ export function LocationRaceSpecific({
                 <div className="sticky top-24 text-center">
                   <DTSIPersonHeroCard
                     countryCode={countryCode}
-                    cryptoStanceGradeElement={<DTSIFormattedLetterGrade person={person} />}
+                    cryptoStanceGrade={DTSIFormattedLetterGrade}
                     isRecommended={isRecommended}
                     person={person}
                     subheader="role"
