+++ src/components/app/pageLocationKeyRaces/us/locationStateSpecific/userLocationRaceInfo.tsx
@@ -106,7 +106,7 @@ function SuspenseUserLocationRaceInfo({
           {recommended && (
             <DTSIPersonHeroCard
               countryCode={countryCode}
-              cryptoStanceGradeElement={<DTSIFormattedLetterGrade person={recommended} />}
+              cryptoStanceGrade={DTSIFormattedLetterGrade}
               isRecommended
               person={recommended}
               subheader="role"
@@ -115,7 +115,7 @@ function SuspenseUserLocationRaceInfo({
           {others.map(person => (
             <DTSIPersonHeroCard
               countryCode={countryCode}
-              cryptoStanceGradeElement={<DTSIFormattedLetterGrade person={person} />}
+              cryptoStanceGrade={DTSIFormattedLetterGrade}
               key={person.id}
               person={person}
               subheader={person.isIncumbent ? 'Incumbent' : 'role'}
